% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{find_env_file}
\alias{find_env_file}
\title{Find a .env file by walking up the directory tree}
\usage{
find_env_file(start_dir = getwd(), filename = ".env")
}
\arguments{
\item{start_dir}{Directory to start searching from (default: getwd())}

\item{filename}{Name of the env file (default: ".env")}
}
\value{
Path to the first matching file or NULL if none found
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
