% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_labarchives.R
\name{la_auth_params}
\alias{la_auth_params}
\title{Build auth parameters for LabArchives API requests}
\usage{
la_auth_params(client, api_method, expires = NULL)
}
\arguments{
\item{client}{LabArchives client}

\item{api_method}{API method name used for signature}

\item{expires}{Optional precomputed expires timestamp}
}
\value{
Named list with akid, expires, sig
}
\description{
Build auth parameters for LabArchives API requests
}
