% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_labarchives.R
\name{la_expires}
\alias{la_expires}
\title{Create an expires timestamp for LabArchives auth}
\usage{
la_expires(offset_seconds = 0)
}
\arguments{
\item{offset_seconds}{Time from now in seconds (default 0)}
}
\value{
Expiration timestamp as a string (milliseconds since epoch)
}
\description{
LabArchives expects an expiry timestamp in milliseconds in the API docs.
}
