% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_labarchives.R
\name{la_user_login}
\alias{la_user_login}
\title{Exchange auth_code/email for user UID}
\usage{
la_user_login(client, auth_code, email)
}
\arguments{
\item{client}{LabArchives client (akid and password)}

\item{auth_code}{Auth code from login redirect}

\item{email}{Email from login redirect}
}
\value{
List with success status and UID
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
