% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_labarchives.R
\name{parse_labarchives_access_info}
\alias{parse_labarchives_access_info}
\title{Parse LabArchives user access info response}
\usage{
parse_labarchives_access_info(xml)
}
\arguments{
\item{xml}{XML document from user_access_info}
}
\value{
List with success flag, user data, or error
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
